first pass of making log files and time logging both feaures
